+++ divi/src/scriptCheck.cpp
@@ -6,7 +6,15 @@
 #include <scriptCheck.h>
 #include <script/sigcache.h>
 #include <Logging.h>
+#include <primitives/transaction.h>
+#include <coins.h>
 
+CScriptCheck::CScriptCheck() : ptxTo(0), nIn(0), nFlags(0), cacheStore(false), error(SCRIPT_ERR_UNKNOWN_ERROR) {}
+
+CScriptCheck::CScriptCheck(const CCoins& txFromIn, const CTransaction& txToIn, unsigned int nInIn, unsigned int nFlagsIn, bool cacheIn) : scriptPubKey(txFromIn.vout[txToIn.vin[nInIn].prevout.n].scriptPubKey),
+                                                                                                                                ptxTo(&txToIn), nIn(nInIn), nFlags(nFlagsIn), cacheStore(cacheIn), error(SCRIPT_ERR_UNKNOWN_ERROR) {}
+
+    
 bool CScriptCheck::operator()()
 {
     const CScript& scriptSig = ptxTo->vin[nIn].scriptSig;
@@ -15,3 +23,16 @@ bool CScriptCheck::operator()()
     }
     return true;
 }
+
+
+void CScriptCheck::swap(CScriptCheck& check)
+{
+    scriptPubKey.swap(check.scriptPubKey);
+    std::swap(ptxTo, check.ptxTo);
+    std::swap(nIn, check.nIn);
+    std::swap(nFlags, check.nFlags);
+    std::swap(cacheStore, check.cacheStore);
+    std::swap(error, check.error);
+}
+
+ScriptError CScriptCheck::GetScriptError() const{return error;}
